Refactoring importance tests
